Migrate build to Autotools.
This is a fairly boilerplate Autotools setup as prescribed in
"Autotools: A Practitioner's Guide to GNU Autoconf, Automake, and
Libtool" by John Calcote.

The ChangeLog was generated from the Git commit log using the following
command:

  git log --date=short --pretty=format:'%ad  %an  <%ae>%n%n%w(72,8,8)%B'

The output was run through sed(1) to convert the indents from spaces to
tabs.