Add option to specify output format
Allow the output format to be specified per the argument to the new `-t'
command-line option.  Its value passed directly to one of the family of
gdk_pixbuf_save() functions, giving it default support for the following
formats: jpeg, png, tiff, ico, and bmp.  Support for other formats might
also be installed.